+++ banditpylib/learners/collaborative_learner/__init__.py
@@ -2,5 +2,4 @@ from .utils import *
 from .lilucb_heur_collaborative import *
 from .collaborative_agent import *
 
-__all__ = ['CollaborativeLearner', 'CollaborativeAgent',
-  'CollaborativeMaster', 'MyCollaborativeMaster']
+__all__ = ['CollaborativeLearner', 'CollaborativeAgent']
